Add files via upload
Some code may be redundant, but this file includes the start page, game over page, and status bar for game page, along with code for certain letters displayed in a 10x30 pixel block. Colors and font sizes may and will be adjusted as required. Spawning of targets in and out have not be included in this file.